Cleaned up patcher styling and code.
